+++ config.php
@@ -3,7 +3,8 @@ if (!defined('PSI_CONFIG_FILE')) {
     /**
      * phpSysInfo version
      */
-    define('PSI_VERSION','3.1.13');
+    define('PSI_VERSION', '3.1.14');
+
     /**
      * phpSysInfo configuration
      */