+++ includes/os/class.Linux.inc.php
@@ -46,11 +46,44 @@ class Linux extends OS
      */
     private function _machine()
     {
-        if ( (CommonFunctions::rfts('/var/log/dmesg', $result, 0, 4096, false) 
-              && preg_match('/^[\s\[\]\.\d]*DMI:\s*(.*)/m', $result, $ar_buf)) 
+        if ( (CommonFunctions::rfts('/var/log/dmesg', $result, 0, 4096, false)
+              && preg_match('/^[\s\[\]\.\d]*DMI:\s*(.*)/m', $result, $ar_buf))
            ||(CommonFunctions::executeProgram('dmesg', '', $result, false)
               && preg_match('/^[\s\[\]\.\d]*DMI:\s*(.*)/m', $result, $ar_buf)) ) {
             $this->sys->setMachine(trim($ar_buf[1]));
+        } else { //data from /sys/devices/virtual/dmi/id/
+            $machine = "";
+            $product = "";
+            $board = "";
+            $bios = "";
+            if (CommonFunctions::rfts('/sys/devices/virtual/dmi/id/board_vendor', $buf, 1, 4096, false) && (trim($buf)!="")) {
+                $machine = trim($buf);
+            }
+            if (CommonFunctions::rfts('/sys/devices/virtual/dmi/id/product_name', $buf, 1, 4096, false) && (trim($buf)!="")) {
+                $product = trim($buf);
+            }
+            if (CommonFunctions::rfts('/sys/devices/virtual/dmi/id/board_name', $buf, 1, 4096, false) && (trim($buf)!="")) {
+                $board = trim($buf);
+            }
+            if (CommonFunctions::rfts('/sys/devices/virtual/dmi/id/bios_version', $buf, 1, 4096, false) && (trim($buf)!="")) {
+                $bios = trim($buf);
+            }
+            if (CommonFunctions::rfts('/sys/devices/virtual/dmi/id/bios_date', $buf, 1, 4096, false) && (trim($buf)!="")) {
+                $bios = trim($bios." ".trim($buf));
+            }
+            if ($product != "") {
+                $machine .= " ".$product;
+            }
+            if ($board != "") {
+                $machine .= "/".$board;
+            }
+            if ($bios != "") {
+                $machine .= ", BIOS ".$bios;
+            }
+
+            if ($machine != "") {
+                $this->sys->setMachine(trim($machine));
+            }
         }
     }
 
@@ -232,6 +265,7 @@ class Linux extends OS
         if (isset($this->_cpu_loads[$cpuline])) {
             return $this->_cpu_loads[$cpuline];
         }
+
         return 0;
     }
 
@@ -491,6 +525,10 @@ class Linux extends OS
                 if (isset($device[6]) && trim($device[6]) != "") {
                     $dev = new HWDevice();
                     $dev->setName(trim($device[6]));
+                    $this->sys->setUsbDevices($dev);
+                } elseif (isset($device[5]) && trim($device[5]) != "") {
+                    $dev = new HWDevice();
+                    $dev->setName("unknown");
                     $this->sys->setUsbDevices($dev);
                 }
             }