+++ CHANGELOG.md
@@ -3,6 +3,17 @@ Changelog of phpSysInfo
 
 http://phpsysinfo.sourceforge.net/
 
+phpSysInfo 3.1.14
+----------------
+
+ - [ADD] FreeIPMI sensor program support
+ - [ADD] IPMIutil sensor program support
+ - [ADD] PowerSoftPlus (EVER) UPS program support
+ - [ADD] Line frequency for the UPS info
+ - [ADD] SENSOR_EVENTS option - show events of sensors
+ - [ADD] HIDE_RAID_DEVICES option for plugins MDStatus and DMRaid - list of RAID devices to hide
+ - [ADD] idash template
+
 phpSysInfo 3.1.13
 ----------------